pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrei-KV",email = "a.v.kovalev9@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"

"sqlalchemy" = ">=2.0.44,<3.0.0"
"langchain-text-splitters" = ">=1.0.0,<2.0.0"
"python-dotenv" = ">=1.2.1,<2.0.0"
"markdown" = ">=3.9,<4.0"
"pypandoc" = ">=1.15,<2.0"
"beautifulsoup4" = ">=4.14.2,<5.0.0"
"pandoc" = ">=2.4,<3.0"
"asyncpg" = ">=0.29.0,<1.0.0"
"redis" = ">=7.0.1,<8.0.0"
"tiktoken" = ">=0.12.0,<0.13.0"
"docx2python" = ">=2.2.0,<3.0.0"
"PyMuPDF" = ">=1.24.0,<2.0.0"
"pytesseract" = ">=0.3.10,<0.4.0"
"Pillow" = ">=10.0.0,<11.0.0"
"nltk" = ">=3.8.1,<4.0.0"
"google-genai" = ">=1.49.0,<2.0.0"
"pandas" = ">=2.3.3,<3.0.0"
"openpyxl" = ">=3.1.5,<4.0.0"
"xlrd" = ">=2.0.2,<3.0.0"
"google-api-python-client" = ">=2.187.0,<3.0.0"
"openai" = ">=1.3.3,<2.0.0"

"transformers" = ">=4.57.1,<5.0.0"


torch = {version = ">=2.9.1,<3.0.0", source = "pytorch-cpu"}

"numpy" = ">=2.3.5,<3.0.0"
"fastapi" = ">=0.121.2,<0.122.0"
"uvicorn" = ">=0.38.0,<0.39.0"
"scipy" = ">=1.16.3,<2.0.0"
"grpcio" = ">=1.76.0,<2.0.0"
"qdrant-client" = ">=1.16.0,<2.0.0"
"sentence-transformers" = ">=5.1.2,<6.0.0"
python-docx = "^1.2.0"
pdfplumber = "^0.11.8"
striprtf = "^0.0.29"
opensearch-py = "^3.1.0"
aiohttp = "^3.13.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"
